feat (47): Removing old home view screen to replace it for Wishlist screen
