--- src/app/src/main/java/com/gabrielbmoro/moviedb/ui/screens/home/HomeUIState.kt
@@ -1,14 +1,0 @@
-package com.gabrielbmoro.moviedb.ui.screens.home
-
-import androidx.paging.PagingData
-import com.gabrielbmoro.moviedb.domain.model.Movie
-import com.gabrielbmoro.moviedb.domain.model.MovieListType
-import kotlinx.coroutines.flow.Flow
-
-data class HomeUIState(
-    val favoriteMovies: List<Movie>? = null,
-    val isLoading: Boolean = false,
-    val selectedMovieType: MovieListType,
-    val areBarsVisible: Boolean = true,
-    val paginatedMovies: Flow<PagingData<Movie>>
-)
