+++ src/app/src/main/java/com/gabrielbmoro/moviedb/ui/screens/wishlist/WishlistUIState.kt
@@ -5,4 +5,5 @@ import com.gabrielbmoro.moviedb.domain.model.Movie
 data class WishlistUIState(
     val favoriteMovies: List<Movie>? = null,
     val isLoading: Boolean = false,
+    val areBarsVisible: Boolean = true,
 )
