+++ src/app/src/main/java/com/gabrielbmoro/moviedb/ui/screens/wishlist/WishlistViewModel.kt
@@ -38,4 +38,10 @@ class WishlistViewModel(
             }
         }
     }
+
+    fun updateScrollPosition(y: Float) {
+        _uiState.update {
+            it.copy(areBarsVisible = y > 0)
+        }
+    }
 }
